refactor(sonic_analysis): drop dead code flagged in PR review
Removed select_clap_window and select_clap_windows from the provider —
the streaming PCM path uses compute_clap_target_starts instead, and the
old helpers had no production callers. Renamed the test file to match.

Also dropped the module-level validate_calibration_freshness() call in
clap_prompts.py; handle_async_init still calls it on provider init, so
the warning still fires.
