+++ music_assistant/providers/sonic_analysis/clap_prompts.py
@@ -142,6 +142,3 @@ def validate_calibration_freshness(
             CALIBRATION_PROMPTS_HASH,
             actual,
         )
-
-
-validate_calibration_freshness()
